\name{pamm-package}
\alias{pamm-package}
\alias{pamm}
\docType{package}
\title{
Power analysis for random effects in mixed models
}

\description{
Description: Simulation functions to assess or explore the power of
a dataset to estimates significant random effects in a mixed model.
The functions are based on the "lme4" package.
}

\details{
  \tabular{ll}{
  Package: \tab pamm\cr
  Type: \tab Package\cr
  Version: \tab 0.9\cr
  Date: \tab 2015-12-10\cr
  License: \tab GPL 2.00\cr
  LazyLoad: \tab yes\cr
  }
}

\author{
Julien Martin
Maintainer:Julien Martin \email{julienmartin@abdn.ac.uk}
}

\references{
Martin, Nussey, Wilson and Reale Submitted Measuring between-individual
variation in reaction norms in field and experimental studies: a power
analysis of random regression models. Methods in Ecology and Evolution.  
}

\keyword{
package 
}

\seealso{
\code{\link{EAMM}}, \code{\link{SSF}}, \code{\link{PAMM}}
}
