% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SSF.R
\name{plot.SSF}
\alias{plot.SSF}
\title{Graphic output of the PAMM function}
\usage{
\method{plot}{SSF}(x, ...)
}
\arguments{
\item{x}{an SSF object}

\item{\dots}{potentially further arguments to pass to methods}
}
\description{
Provide graphic interpretation of the simulation results
}
\examples{
\dontrun{
   oursSSF <- SSF(50,100,10,c(0.4,0.1,0.6,0))
   plot(oursSSF)
   }


}
