% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamm_package.R
\docType{package}
\name{pamm-package}
\alias{pamm}
\alias{pamm-package}
\title{pamm: Power Analysis for Random Effects in Mixed Models}
\description{
Simulation functions to assess or explore the power of a dataset to estimates significant random effects (intercept or slope) in a mixed model. The functions are based on the "lme4" and "lmerTest" packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JulienGAMartin/pamm_R}
  \item Report bugs at \url{https://github.com/JulienGAMartin/pamm_R/issues}
}

}
\author{
\strong{Maintainer}: Julien Martin \email{julien.martin@uottawa.ca}

}
\keyword{package}
